type-c-service: Move most of `utils` into `type-c-interface`
Move these functions because they're generally useful and not specific
to `type-c-service`.
