+++ examples/std/src/lib/type_c/mock_controller.rs
@@ -21,8 +21,8 @@ use type_c_interface::control::type_c::TypeCStateMachineState;
 use type_c_interface::control::usb::UsbControlConfig;
 use type_c_interface::control::vdm::{AttnVdm, OtherVdm, SendVdm};
 use type_c_interface::port::event::PortEventBitfield;
+use type_c_interface::util::power_capability_from_current;
 use type_c_service::controller::state::SharedState;
-use type_c_service::util::power_capability_from_current;
 
 pub struct ControllerState {
     events: Signal<GlobalRawMutex, PortEventBitfield>,
