+++ type-c-interface/src/lib.rs
@@ -5,3 +5,4 @@ pub mod controller;
 pub mod port;
 pub mod service;
 pub mod ucsi;
+pub mod util;
