+++ type-c-service/tests/power.rs
@@ -12,8 +12,9 @@ use power_policy_interface::{
 use type_c_interface::{
     control::pd::PortStatus,
     port::event::{PortEvent, PortStatusEventBitfield},
+    util::POWER_CAPABILITY_5V_1A5,
 };
-use type_c_service::{controller::event::Event, util::POWER_CAPABILITY_5V_1A5};
+use type_c_service::controller::event::Event;
 
 use crate::common::{
     DEFAULT_PER_CALL_TIMEOUT, DEFAULT_TEST_DURATION, PowerPolicyServiceReceiver, Test, TestPort, TypeCServiceReceiver,
